cleaning up code - wrapping state for add semester
